Rename the test and extend from AcceptanceTestBase
Add tests for starting the extensions and stopping them
